+++ CumulusMX/Stations/EcowittApi.cs
@@ -2525,13 +2525,13 @@ namespace CumulusMX.Stations
 		}
 
 
-		internal string GetCurrentCameraImageUrl(string defaultUrl, int camIndex, CancellationToken token)
+		internal string GetCurrentCameraImageUrl(string defaultUrl, string mac, CancellationToken token)
 		{
 			// Doc: https://doc.ecowitt.net/web/#/apiv3en?page_id=17
 
 			cumulus.LogMessage("API.GetCurrentCameraImageUrl: Get Ecowitt Current Camera Data");
 
-			if (string.IsNullOrEmpty(cumulus.EcowittApplicationKey) || string.IsNullOrEmpty(cumulus.EcowittUserApiKey) || string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress[camIndex]))
+			if (string.IsNullOrEmpty(cumulus.EcowittApplicationKey) || string.IsNullOrEmpty(cumulus.EcowittUserApiKey) || string.IsNullOrEmpty(mac))
 			{
 				cumulus.LogWarningMessage("API.GetCurrentCameraImageUrl: Missing Ecowitt API data in the configuration, aborting process");
 				return defaultUrl;
@@ -2557,7 +2557,7 @@ namespace CumulusMX.Stations
 
 			sb.Append($"application_key={cumulus.EcowittApplicationKey}");
 			sb.Append($"&api_key={cumulus.EcowittUserApiKey}");
-			sb.Append($"&mac={cumulus.EcowittCameraMacAddress[camIndex]}");
+			sb.Append($"&mac={mac}");
 			sb.Append("&call_back=camera");
 
 			var url = sb.ToString();
@@ -2660,7 +2660,7 @@ namespace CumulusMX.Stations
 			}
 		}
 
-		internal string GetLastCameraVideoUrl(string defaultUrl, int camIndex, CancellationToken token)
+		internal string GetLastCameraVideoUrl(string defaultUrl, string mac, CancellationToken token)
 		{
 			// Doc: https://doc.ecowitt.net/web/#/apiv3en?page_id=19
 
@@ -2681,7 +2681,7 @@ namespace CumulusMX.Stations
 						*/
 			cumulus.LogMessage("API.GetLastCameraVideoUrl: Get Ecowitt Last Camera Video");
 
-			if (string.IsNullOrEmpty(cumulus.EcowittApplicationKey) || string.IsNullOrEmpty(cumulus.EcowittUserApiKey) || string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress[camIndex]))
+			if (string.IsNullOrEmpty(cumulus.EcowittApplicationKey) || string.IsNullOrEmpty(cumulus.EcowittUserApiKey) || string.IsNullOrEmpty(mac))
 			{
 				cumulus.LogWarningMessage("API.GetLastCameraVideoUrl: Missing Ecowitt API data in the configuration, aborting process");
 				return defaultUrl;
@@ -2709,7 +2709,7 @@ namespace CumulusMX.Stations
 
 			sb.Append($"application_key={cumulus.EcowittApplicationKey}");
 			sb.Append($"&api_key={cumulus.EcowittUserApiKey}");
-			sb.Append($"&mac={cumulus.EcowittCameraMacAddress[camIndex]}");
+			sb.Append($"&mac={mac}");
 			sb.Append($"&start_date={start:yyyy-MM-dd'%20'HH:mm:ss}");
 			sb.Append($"&end_date={end:yyyy-MM-dd'%20'HH:mm:ss}");
 			sb.Append("&call_back=camera.video");
@@ -2939,7 +2939,10 @@ namespace CumulusMX.Stations
 						if (stn.type == 2 && CheckCamera)
 						{
 							// we have a camera
-							cumulus.EcowittCameraMacAddress.Add(stn.mac);
+							if (!cumulus.EcowittCameraMacAddress.Contains(stn.mac))
+							{
+								cumulus.EcowittCameraMacAddress.Add(stn.mac);
+							}
 							cumulus.LogDebugMessage($"API.GetStationList: Found Camera name={stn.name ?? "-"}, vers={stn.stationtype ?? "-"}, mac={stn.mac}");
 						}
 						else if (stn.type == 1 && stn.mac.Equals(macAddress, StringComparison.CurrentCultureIgnoreCase))