+++ CumulusMX/Stations/EcowittCloudStation.cs
@@ -261,11 +261,11 @@ namespace CumulusMX.Stations
 			StartLoop();
 		}
 
-		public override string GetEcowittCameraUrl(int camIndex)
+		public override string GetEcowittCameraUrl(string mac)
 		{
 			if (cumulus.ExtraSensorUseCamera ^ mainStation)
 			{
-				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress[camIndex]))
+				if (string.IsNullOrEmpty(mac))
 				{
 					cumulus.LogWarningMessage("GetEcowittCameraUrl: Warning - URL requested, but no camera MAC address is configured");
 				}
@@ -273,8 +273,8 @@ namespace CumulusMX.Stations
 				{
 					try
 					{
-						EcowittCameraUrl[camIndex] = ecowittApi.GetCurrentCameraImageUrl(EcowittCameraUrl[camIndex], camIndex, Program.ExitSystemToken);
-						return EcowittCameraUrl[camIndex];
+						EcowittCameraUrl[mac] = ecowittApi.GetCurrentCameraImageUrl(EcowittCameraUrl[mac], mac, Program.ExitSystemToken);
+						return EcowittCameraUrl[mac];
 					}
 					catch (Exception ex)
 					{
@@ -286,11 +286,11 @@ namespace CumulusMX.Stations
 			return string.Empty;
 		}
 
-		public override string GetEcowittVideoUrl(int camIndex)
+		public override string GetEcowittVideoUrl(string mac)
 		{
 			if (cumulus.ExtraSensorUseCamera ^ mainStation)
 			{
-				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress[camIndex]))
+				if (string.IsNullOrEmpty(mac))
 				{
 					cumulus.LogWarningMessage("GetEcowittCameraUrl: Warning - URL requested, but no camera MAC address is configured");
 				}
@@ -298,8 +298,8 @@ namespace CumulusMX.Stations
 				{
 					try
 					{
-						EcowittVideoUrl[camIndex] = ecowittApi.GetLastCameraVideoUrl(EcowittVideoUrl[camIndex], Program.ExitSystemToken);
-						return EcowittVideoUrl[camIndex];
+						EcowittVideoUrl[mac] = ecowittApi.GetLastCameraVideoUrl(EcowittVideoUrl[mac], mac, Program.ExitSystemToken);
+						return EcowittVideoUrl[mac];
 					}
 					catch (Exception ex)
 					{
@@ -617,17 +617,19 @@ namespace CumulusMX.Stations
 				}
 
 				// === Camera ===
+				/*
 				try
 				{
 					if (data.camera != null && data.camera.photo != null)
 					{
-						EcowittCameraUrl = data.camera.photo.url;
+						EcowittCameraUrl[0] = data.camera.photo.url;
 					}
 				}
 				catch (Exception ex)
 				{
 					cumulus.LogErrorMessage($"ProcessCurrentData: Error in Camera data - {ex.Message}");
 				}
+				*/
 
 				if (cumulus.StationOptions.CalculateSLP)
 				{