+++ CHANGELOG.md
@@ -14,8 +14,8 @@ Alternatively view it [online on GitHub](https://github.com/cumulusmx/CumulusMX/
 
 ## Changed
 
-- If you have more than one Ecowitt web camera, the web tags `<#EcowittCameraUrl>` and `<#EcowittVideoUrl>` now accept a parameter of `cam=N` (eg. `<#EcowittCameraUrl cam=2>`) to return the zero relative camera from your account.<br>
-	If omitted the parameter defaults to 0 (zero) and returns the first (or only camera)
+- If you have more than one Ecowitt web camera, the web tags `<#EcowittCameraUrl>` and `<#EcowittVideoUrl>` now accept a parameter of `mac=1A:2B:3C:4D:5E:6F` (eg. `<#EcowittCameraUrl mac=3C:8A:1F:B3:2B:A2>`) to return the matching camera from your account.<br>
+	If omitted the parameter defaults to the first (or only camera) found on your account
 
 ### Fixed